+++ modules/elementary_functions/src/cpp/MatrixMultiplication.cpp
@@ -406,7 +406,8 @@ matrixMultiplication(ArrayOf& A, ArrayOf& B, bool& needToOverload)
             }
             A.promoteType(B.getDataClass());
             return matrixMultiplication(A, B, needToOverload);
-        } else if (isIntegerA && isIntegerB) {
+        }
+        if (isIntegerA && isIntegerB) {
             if (A.getDataClass() != B.getDataClass()) {
                 needToOverload = true;
             } else {