+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 # 0.3.12 (2019-12-XX)
 
+* sort builtin: sort double, single, integers, cell and strings.
+
 * diag builtin: Get diagonal elements of matrix or create diagonal matrix.
 
 * Continous Integration tools for external modules (see module skeleton example).